+++ MovieRatingShared/PlayerMovieScore.cs
@@ -4,5 +4,5 @@ public class PlayerMovieScore
 {
     public required int MovieIndex { get; init; }
     public required int MovieScore { get; init; }
-    public required int MovieRating { get; init; }
+    public required double MovieRating { get; init; }
 }
